# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "pyKorf",
    "version": "0.18.0"
  },
  "paths": {
...
    "/api/preferences/sp-overrides/add": {
      "post": {
        "tags": [
          "preferences"
        ],
        "summary": "Add Sp Override",
        "description": "Add a new SharePoint override.",
        "operationId": "addSpOverride",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddSpOverrideRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OkResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/preferences/sp-overrides/edit": {
      "post": {
        "tags": [
          "preferences"
        ],
        "summary": "Edit Sp Override",
        "description": "Edit an existing SharePoint override.",
        "operationId": "editSpOverride",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EditSpOverrideRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OkResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/preferences/sp-overrides/delete": {
      "post": {
        "tags": [
          "preferences"
        ],
        "summary": "Delete Sp Override",
        "description": "Delete a SharePoint override.",
        "operationId": "deleteSpOverride",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DeleteSpOverrideRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OkResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/preferences/skip-sp": {
      "post": {
        "tags": [
          "preferences"
        ],
        "summary": "Set Skip Sp",
        "description": "Toggle skip SharePoint override validation.",
        "operationId": "setSkipSp",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SetSkipSpRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OkResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/preferences/license": {
      "post": {
        "tags": [
          "preferences"
        ],
        "summary": "Set License",
        "description": "Validate and save a license key.",
        "operationId": "setLicense",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SetLicenseKeyRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LicenseValidationResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/preferences/doc-register": {
      "post": {
        "tags": [
          "preferences"
        ],
        "summary": "Set Doc Register Config",
        "description": "Save Document Register configuration and rebuild DB if Excel path changed.",
        "operationId": "setDocRegisterConfigFromPreferences",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SetDocRegisterConfigRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DocRegisterRebuildResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AddSpOverrideRequest": {
        "properties": {
          "local_path": {
            "type": "string",
            "title": "Local Path"
          },
          "sp_url": {
            "type": "string",
            "title": "Sp Url"
          }
        },
        "type": "object",
        "required": [
          "local_path",
          "sp_url"
        ],
        "title": "AddSpOverrideRequest"
      },
...
      "DeleteSpOverrideRequest": {
        "properties": {
          "local_path": {
            "type": "string",
            "title": "Local Path"
          }
        },
        "type": "object",
        "required": [
          "local_path"
        ],
        "title": "DeleteSpOverrideRequest"
      },
...
      "DocRegisterRebuildResponse": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success",
            "default": false
          },
          "message": {
            "type": "string",
            "title": "Message",
            "default": ""
          },
          "stats": {
            "additionalProperties": true,
            "type": "object",
            "title": "Stats",
            "default": {}
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          }
        },
        "type": "object",
        "title": "DocRegisterRebuildResponse"
      },
...
      "EditSpOverrideRequest": {
        "properties": {
          "original_local_path": {
            "type": "string",
            "title": "Original Local Path"
          },
          "local_path": {
            "type": "string",
            "title": "Local Path"
          },
          "sp_url": {
            "type": "string",
            "title": "Sp Url"
          }
        },
        "type": "object",
        "required": [
          "original_local_path",
          "local_path",
          "sp_url"
        ],
        "title": "EditSpOverrideRequest"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "LicenseValidationResponse": {
        "properties": {
          "valid": {
            "type": "boolean",
            "title": "Valid"
          },
          "expiry": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Expiry"
          },
          "error": {
            "type": "string",
            "title": "Error",
            "default": ""
          }
        },
        "type": "object",
        "required": [
          "valid"
        ],
        "title": "LicenseValidationResponse"
      },
...
      "OkResponse": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success"
          },
          "message": {
            "type": "string",
            "title": "Message",
            "default": ""
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          }
        },
        "type": "object",
        "required": [
          "success"
        ],
        "title": "OkResponse",
        "description": "Generic success/failure response for simple write operations."
      },
...
      "SetDocRegisterConfigRequest": {
        "properties": {
          "excel_path": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Excel Path"
          },
          "sp_site_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sp Site Url"
          }
        },
        "type": "object",
        "title": "SetDocRegisterConfigRequest"
      },
      "SetLicenseKeyRequest": {
        "properties": {
          "license_key": {
            "type": "string",
            "title": "License Key"
          }
        },
        "type": "object",
        "required": [
          "license_key"
        ],
        "title": "SetLicenseKeyRequest"
      },
...
      "SetSkipSpRequest": {
        "properties": {
          "skip": {
            "type": "boolean",
            "title": "Skip"
          }
        },
        "type": "object",
        "required": [
          "skip"
        ],
        "title": "SetSkipSpRequest"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}